requirements.txt
sounddevice
numpy
pyaudio
python-dotenv
google-cloud-speech
openai
elevenlabs